+++ dec_13/src/main.rs
@@ -5,20 +5,20 @@ fn main() {
 	println!("{} {}", sol1, sol2);
 }
 
-fn solve1(input: &str) -> usize {
+fn solve1(input: &str) -> u64 {
 	let linebreak = input.chars().position(|c| c == '\n').unwrap();
-	let time_stamp = input[..linebreak].parse::<usize>().unwrap();
+	let time_stamp = input[..linebreak].parse::<u64>().unwrap();
 	let (bus, min_wait) = input[(linebreak + 1)..]
 		.trim()
 		.split(',')
-		.filter_map(|s| s.parse::<usize>().ok())
+		.filter_map(|s| s.parse::<u64>().ok())
 		.map(|b| (b, b - time_stamp % b))
 		.min_by_key(|(_, b)| *b)
 		.unwrap();
 	bus * min_wait
 }
 
-fn solve2(input: &str) -> usize {
+fn solve2(input: &str) -> u64 {
 	let linebreak = input.chars().position(|c| c == '\n').unwrap();
 	let mut buses = input[(linebreak + 1)..]
 		.trim()
@@ -26,9 +26,9 @@ fn solve2(input: &str) -> usize {
 		.enumerate()
 		.map(|(i, s)| (i, s.parse()))
 		.filter(|(_, x)| x.is_ok())
-		.map(|(a, b)| (a, b.unwrap()))
+		.map(|(a, b)| (a as u64, b.unwrap()))
 		.map(|(a, b)| ((16 * b - a) % b, b))
-		.collect::<Vec<(usize, usize)>>();
+		.collect::<Vec<(u64, u64)>>();
 	buses.sort_unstable_by_key(|(_, a)| *a);
 	let (lcm, res) = buses
 		.iter()
@@ -45,7 +45,7 @@ fn solve2(input: &str) -> usize {
 	res
 }
 
-fn find_mod(x1: usize, r1: usize, x2: usize, r2: usize) -> (usize, usize) {
+fn find_mod(x1: u64, r1: u64, x2: u64, r2: u64) -> (u64, u64) {
 	let lcm = x1 * x2;
 	let r2 = (r2 + x2 * 16) % x2;
 	let r1 = (r1 + x1 * 16) % x1;
